Tkinter ML/requirements.txt
absl-py==0.9.0
altgraph==0.17
astor==0.8.1
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.2
pefile==2019.4.18
Pillow==7.0.0
protobuf==3.6.0
PyInstaller==3.6
pywin32-ctypes==0.2.0
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==1.0.0
wrapt==1.12.1
